requirements.txt
asgiref==3.2.10
Django==3.0.7
django-cors-headers==3.4.0
django-environ==0.4.5
django-summernote==0.8.11.6
django-taggit==1.3.0
djangorestframework==3.11.0
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
asgiref==3.2.10
Django==3.0.7
django-cors-headers==3.4.0
django-environ==0.4.5
django-summernote==0.8.11.6
django-taggit==1.3.0
djangorestframework==3.11.0
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
